Changed Fetch strategy to and added track search
Now instead of usual fetch services there is a RTK Query because it's easy to use and to scale the app with it.
With new fetching strategy, the track searching page and adding already uploaded tracks to album feature was added.
Edited some minor bugs and optimized playlist to prevent lags from lerge tracks amount.